+++ app/ante/eth.go
@@ -237,13 +237,18 @@ func (nvd EthNonceVerificationDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, 
 // EthGasConsumeDecorator validates enough intrinsic gas for the transaction and
 // gas consumption.
 type EthGasConsumeDecorator struct {
-	evmKeeper EVMKeeper
+	evmKeeper    EVMKeeper
+	maxGasWanted uint64
 }
 
 // NewEthGasConsumeDecorator creates a new EthGasConsumeDecorator
-func NewEthGasConsumeDecorator(evmKeeper EVMKeeper) EthGasConsumeDecorator {
+func NewEthGasConsumeDecorator(
+	evmKeeper EVMKeeper,
+	maxGasWanted uint64,
+) EthGasConsumeDecorator {
 	return EthGasConsumeDecorator{
-		evmKeeper: evmKeeper,
+		evmKeeper,
+		maxGasWanted,
 	}
 }
 
@@ -293,8 +298,8 @@ func (egcd EthGasConsumeDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simula
 
 		if ctx.IsCheckTx() {
 			// We can't trust the tx gas limit, because we'll refund the unused gas.
-			if txData.GetGas() > MaxTxGasWanted {
-				gasWanted += MaxTxGasWanted
+			if txData.GetGas() > egcd.maxGasWanted {
+				gasWanted += egcd.maxGasWanted
 			} else {
 				gasWanted += txData.GetGas()
 			}